// src/gql/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
    ID: { input: string; output: string };
    String: { input: string; output: string };
    Boolean: { input: boolean; output: boolean };
    Int: { input: number; output: number };
    Float: { input: number; output: number };
};

export type Continent = {
    __typename?: 'Continent';
    code: Scalars['ID']['output'];
    countries: Array<Country>;
    name: Scalars['String']['output'];
};

export type ContinentFilterInput = {
    code: InputMaybe<StringQueryOperatorInput>;
};

export type Country = {
    __typename?: 'Country';
    awsRegion: Scalars['String']['output'];
    capital: Maybe<Scalars['String']['output']>;
    code: Scalars['ID']['output'];
    continent: Continent;
    currencies: Array<Scalars['String']['output']>;
    currency: Maybe<Scalars['String']['output']>;
    emoji: Scalars['String']['output'];
    emojiU: Scalars['String']['output'];
    languages: Array<Language>;
    name: Scalars['String']['output'];
    native: Scalars['String']['output'];
    phone: Scalars['String']['output'];
    phones: Array<Scalars['String']['output']>;
    states: Array<State>;
    subdivisions: Array<Subdivision>;
};

export type CountryNameArgs = {
    lang: InputMaybe<Scalars['String']['input']>;
};

export type CountryFilterInput = {
    code: InputMaybe<StringQueryOperatorInput>;
    continent: InputMaybe<StringQueryOperatorInput>;
    currency: InputMaybe<StringQueryOperatorInput>;
    name: InputMaybe<StringQueryOperatorInput>;
};

export type Language = {
    __typename?: 'Language';
    code: Scalars['ID']['output'];
    name: Scalars['String']['output'];
    native: Scalars['String']['output'];
    rtl: Scalars['Boolean']['output'];
};
// ...
export type State = {
    __typename?: 'State';
    code: Maybe<Scalars['String']['output']>;
    country: Country;
    name: Scalars['String']['output'];
};

export type StringQueryOperatorInput = {
    eq: InputMaybe<Scalars['String']['input']>;
    in: InputMaybe<Array<Scalars['String']['input']>>;
    ne: InputMaybe<Scalars['String']['input']>;
    nin: InputMaybe<Array<Scalars['String']['input']>>;
    regex: InputMaybe<Scalars['String']['input']>;
};

export type Subdivision = {
    __typename?: 'Subdivision';
    code: Scalars['ID']['output'];
    emoji: Maybe<Scalars['String']['output']>;
    name: Scalars['String']['output'];
};